refactor: renaming app
