+++ backend/backend/apps/authentication/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class CoreAuthConfig(AppConfig):
+class AuthenticationConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "backend.apps.core_auth"
+    name = "backend.apps.authentication"
