+++ backend/backend/apps/bookings/views.py
@@ -6,7 +6,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from backend.apps.core_auth.auth import APIKeyHeaderAuthentication
+from backend.apps.authentication.auth import APIKeyHeaderAuthentication
 
 
 class FreeBookingTime(APIView):
